refactor: extract Layout component
